+++ scripts/macros/spells/conjureCelestial.js
@@ -24,5 +24,5 @@ export async function conjureCelestial({speaker, actor, token, character, item, 
     };
     let animation = chris.getConfiguration(workflow.item, 'animation') ?? 'celestial';
     if (chris.jb2aCheck() != 'patreon' || !chris.aseCheck()) animation = 'none';
-    await summons.spawn(sourceActors, updates, 3600, workflow.item, undefined, undefined, 90, workflow.token, animation);
+    await summons.spawn(sourceActors, updates, 3600, workflow.item, workflow.token, workflow.item.system?.range?.value, {'spawnAnimation': animation});
 }