+++ scripts/macros/spells/conjureMinorElementals.js
@@ -35,5 +35,5 @@ export async function conjureMinorElementals({speaker, actor, token, character, 
     };
     let animation = chris.getConfiguration(workflow.item, 'animation') ?? 'default';
     if (chris.jb2aCheck() != 'patreon' || !chris.aseCheck()) animation = 'none';
-    await summons.spawn(sourceActors, updates, 3600, workflow.item, undefined, undefined, 90, workflow.token, animation);
+    await summons.spawn(sourceActors, updates, 3600, workflow.item, workflow.token, workflow.item.system?.range?.value, {'spawnAnimation': animation});
 }