+++ scripts/macros/spells/findGreaterSteed.js
@@ -70,7 +70,7 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     };
     let animation = chris.getConfiguration(workflow.item, 'animation-' + creatureType) ?? defaultAnimations[creatureType];
     if (chris.jb2aCheck() != 'patreon' || !chris.aseCheck()) animation = 'none';
-    await summons.spawn(sourceActor, updates, 86400, workflow.item, undefined, undefined, 30, workflow.token, animation);
+    await summons.spawn(sourceActor, updates, 86400, workflow.item, workflow.token, workflow.item.system?.range?.value, {'spawnAnimation': animation});
     let effect = chris.findEffect(workflow.actor, workflow.item.name);
     await chris.updateEffect(effect, updates2);
 }