+++ CHANGELOG.md
@@ -1,10 +1,17 @@
+# 0.10.76 Release:
+## Update Notes:
+- Added pocket dimension feature to find familiar. You can use the new "Pocket Dimension" feature on the actor to dismiss or summon the familair. This works between scenes and will delete left behind familiar tokens. Deleted familiar tokens will also be tracked. Health will be retained.
+- Added module and settings check on client load in.
+- Added Midi-QOL Actor & Item on use macro settings to apply recommended settings feature.
+- Changed format of variables passed in chrisPremades.summons.spawn, this should only matter if you use this function for custom summons.
+  
 # 0.10.75 Release:
 ## Bug Fixes:
 - Healing Spirit
 - Favored Foe
-##New Enhanced Animations:
+## New Enhanced Animations:
 - Healing Spirit (Thanks Eskiemoh)
-##Update Notes:
+## Update Notes:
 - Requires D&D system version 3.1.2 or greater. Core concentration and damage application is not supported yet.
 - Support for core concentration and damage application will be in a future release.